Reference

Base.collect — Method

collect(d)

collect the block array d to the calling process, returning a block array

Base.convert — Method

convert(Array,d)

collect the block array d and convert it to a Julia array

Base.eltype — Method

eltype(R)

returns the element type of R

Base.size — Method

size(R)

returns the size of R

Distributed.nprocs — Method

nprocs(R)

returns the number of Julia processes associated with R

Distributed.procs — Method

procs(R)

returns Julia processes associated with R

DistributedArrays.localindices — Method

localindices(R)

returns the indicies of R that are local to the calling process

DistributedJets.blockmap — Method

blockmap(A)

returns map between process id and block indicies

DistributedJets.localblockindices — Method

localblockindices(R)

returns the block indicies that are local to the calling process

Jets.getblock! — Method

getblock!(d, i, x)

get block i::Int and put it into x

Jets.getblock — Method

getblock(d,i)

retrive block i from a distributed block operator

Jets.setblock! — Method

setblock!(d,i,x)

put x into block i::Int

Index

Base.collect

Base.convert

Base.eltype

Base.size

Distributed.nprocs

Distributed.procs

DistributedArrays.localindices

DistributedJets.blockmap

DistributedJets.localblockindices

Jets.getblock

Jets.getblock!

Jets.setblock!